move lifecycle script tests
